Massive simplification of Pentaho configuration
Configurations are deduced wherever possible to reduce configuration points. Since we're going to be in a controlled environment, we have the luxury of limiting what can or can't be configured by way of this chart. If more flexibility is required later on, this can be (re)developed as necessary. For now, this is enough.