Added support for automatic enterprise detection
We use the presence of licenses to identify if a specific chart should run in community or enterprise mode. The presence of licenses generally speaking suggests that enterprise mode should be enabled. "Community licenses" generally speaking aren't a thing b/c why put a license on something that's free to use?