Always boot a Minio container for internal content
This is now needed for clustering. Furthermore, if the main content store is an embedded Minio instance, this implementation will use that same instance for the internal content store. This is fully backwards compatible with Alfresco or Box deployments.

TODO: Support storing the internal content in *actual* (i.e. Amazon) S3 ...